+++ mitogen/imports/__init__.py
@@ -1,0 +1,7 @@
+# SPDX-FileCopyrightText: 2025 Mitogen authors <https://github.com/mitogen-hq>
+# SPDX-License-Identifier: MIT
+# !mitogen: minify_safe
+
+from mitogen.imports._dis import (
+    scan_code_imports as scan_code_imports,
+)